Clean some code up; force GC on every cycle
